requirements.txt
Django==5.0.2
djangorestframework==3.14.0
celery==5.3.6
redis==5.0.1
pandas==2.2.0
numpy==1.26.3
psycopg2-binary==2.9.9
openpyxl==3.1.2
kombu==5.3.5
amqp==5.2.0
six==1.16.0
python-dateutil==2.8.2
pytz==2023.3.post1
gunicorn==21.2.0